fix: check for compatible versions only
Instead of checking for ^{major}.0.0, we're checking for ^{version} now,
which should reflect the actually intended behavior. Aside of that, if
no matching version is found, we're accepting the publish, because
without that behavior you couldn't publish any incompatible updates
anymore.

Aside of that, the check was duplicated, it happened in both the API and
the index. This duplicate was removed, to make sure we don't have
conflicting checks here again in the future.